inventory: Add Position decorator interface (#189)
This decorator interface is used for holding the position of an
inventory item when there are multiple of the same type.

The alternative interface is Decorator.Slot.  The definition of a slot,
as applicable to this context, is 'a narrow opening or groove'.  An
example in a computer is a PCIe slot, so the Slot interface would be
appropriate there.

In other situations for ordering things that don't fit the slot
definition this Position interface would be more appropriate.

In the long run it probably doesn't matter that much as long as it's
consistent in the D-bus objects that the code looking for the interface
uses.

Change-Id: I1fbf9763c3854f5092081a8780d2a56ef0c23c68

Signed-off-by: Matt Spinler <spinler@us.ibm.com>